Create main.py
This is the main program.
To run this program, make sure you have Tkinter and cryptography libraries installed 
(pip install tkinter cryptography). When you run the program, it will generate an encryption key file called key.key and a password file called passwords.txt in the same directory.

The program provides a simple GUI where you can enter a username and password. When you click "Save Password," it will encrypt the password using the encryption key, and save the encrypted password along with the username to the text file. When you click "Retrieve Password," it will retrieve the encrypted password from the text file based on the username and decrypt it using the encryption key.